POC update systray menu dynamically
the QAction objects passed to menu.addAction are passed by reference, so if the QAction object is deleted then the menu item also is removed.
I.E. I was declaring in method scope but then outside the method the object didn't exist and so wouldn't show in the menu.
As we don't know how many menu entries we will have and we can't dynamically declare new variable, we populate a list with these objects, the list is then assigned as an attribute of the main window object instance